refactor: Refactoring row items into its own file
